+++ tests/directives/test_autocdata.py
@@ -27,9 +27,30 @@ class TestAutoCData:
         float b
         """
 
+    unknown_file_level_variable_type = """\
+        static unknown_type *unknown_type_var
+        A variable with unknown type
+        This one we can parse the tokens and try to replace clang's int
+        usage with a stab at the underlying type.  We can't take this token
+        for token as sphinx is too strict at parsing and will assume that
+        MAYBE_CONST is the type."""
+
+    unknown_file_level_array_type = """\
+        int unknown_array_type_var[24]
+        A an array variable with an unknown type.
+        For whatever reason clang will come back with no extent on this so
+        we have to fall back to this being treated as an int"""
+
+    unknown_extern_file_variable = """\
+        extern unknown_type *unknown_extern_type_var
+        Unknown extern type"""
+
     doc_data = [
         ("variables.c::file_level_variable", file_level_variable),
         ("example.c::inline_struct_variable", inline_struct_variable),
+        ("variables.c::unknown_type_var", unknown_file_level_variable_type),
+        ("variables.c::unknown_array_type_var", unknown_file_level_array_type),
+        ("variables.c::unknown_extern_type_var", unknown_extern_file_variable),
     ]
 
     @pytest.mark.parametrize("variable, expected_doc", doc_data)