+++ docs/conf.py
@@ -25,9 +25,9 @@ copyright = u'2019-2020, Nick'
 author = u'Nick'
 
 # The short X.Y version
-version = u'0.3'
+version = u'0.4'
 # The full version, including alpha/beta/rc tags
-release = u'0.3.1'
+release = u'0.4.0'
 
 
 # -- General configuration ---------------------------------------------------